Revamp registration page design and functionality and add Login component
A major overhaul to the registration page was carried out. This involved modifying the css for a cleaner look and advanced functionality. New Angular components for the student dashboard were also created, along with the implementation of environment files for production and development environments. The agreedTerms field was added in the registration form for user agreement confirmation.

Signed-off-by: Geani-Gabriel Dumitrache <dumitrachegeanigabriel@gmail.com>